+++ custom_components/wait_for_wolt/const.py
@@ -14,10 +14,10 @@ REFRESH_URL = "https://authentication.wolt.com/v1/wauth2/access_token"
 # Updated endpoints based on the current Wolt web client
 ACTIVE_ORDERS_URL = "https://consumer-api.wolt.com/order-xp/web/v1/pages/orders"
 ORDER_DETAILS_URL = (
-    "https://restaurant-api.wolt.com/v2/order_details/purchase_tracking?purchase_id={}"
-)
-ORDER_DETAILS_PATH_URL = (
     "https://restaurant-api.wolt.com/v2/order_details/purchase_tracking/{}"
+)
+ORDER_DETAILS_FALLBACK_URL = (
+    "https://restaurant-api.wolt.com/v2/order_details/purchase_tracking?purchase_id={}"
 )
 VENUE_CONTENT_URL = "https://consumer-api.wolt.com/order-xp/web/v1/venue/slug/{}/dynamic/?selected_delivery_method=homedelivery"